requirements.txt
PyMuPDF==1.23.5
pytesseract==0.3.10
Pillow==10.0.0
torch==2.2.1
transformers==4.33.2
